requirements.txt
asgiref==3.8.1
certifi==2021.10.8
charset-normalizer==2.0.10
diff-match-patch==20241021
Django==5.2
django-appconf==1.1.0
django-cleanup==5.2.0
django-crispy-forms==2.3
django-etc==1.4.0
django-hitcount==1.3.5
django-meta==2.4.2
django-modeltranslation==0.19.13
django-parler==2.3
django-recaptcha==3.0.0
django-tinymce==4.1.0
idna==3.3
Pillow==9.0.0
polib==1.2.0
psycopg2-binary==2.9.10
PyMySQL==1.0.2
python-decouple==3.5
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.4.2
tablib==3.8.0
text-unidecode==1.3
typing_extensions==4.13.1
urllib3==1.26.8
